register database public function
